+++ src/main/java/com/buransky/max7219/register/DecodeModeRegister.java
@@ -6,24 +6,24 @@ import com.buransky.max7219.Register;
  * Table 4. Decode-Mode Register Examples (Address (Hex) = 0xX9)
  */
 public enum DecodeModeRegister implements Register {
-    NoDecode((byte)0x00),
-    CodeBDecodeFor0((byte)0x01),
-    CodeBDecodeFor3to0((byte)0x0F),
-    CodeBDecodeFor7to0((byte)0xFF);
+    NoDecode((short)0x00),
+    CodeBDecodeFor0((short)0x01),
+    CodeBDecodeFor3to0((short)0x0F),
+    CodeBDecodeFor7to0((short)0xFF);
 
     private static final RegisterAddress registerAddress = RegisterAddress.DecodeMode;
-    private final byte data;
+    private final short data;
 
-    DecodeModeRegister(final byte data) {
+    DecodeModeRegister(final short data) {
         this.data = data;
     }
 
     @Override
-    public byte getAddress() {
+    public short getAddress() {
         return registerAddress.getAddress();
     }
 
-    public byte getData() {
+    public short getData() {
         return data;
     }
 }
