+++ src/main/java/com/buransky/max7219/register/DigitRegister.java
@@ -8,9 +8,9 @@ import static com.google.common.base.Preconditions.checkArgument;
 public class DigitRegister implements Register {
     public static final RegisterAddress[] DIGITS = { Digit0, Digit1, Digit2, Digit3, Digit4, Digit5, Digit6, Digit7 };
     private final RegisterAddress registerAddress;
-    private final byte data;
+    private final short data;
 
-    public DigitRegister(final RegisterAddress registerAddress, final byte data) {
+    public DigitRegister(final RegisterAddress registerAddress, final short data) {
         checkArgument(registerAddress.getAddress() >= Digit0.getAddress());
         checkArgument(registerAddress.getAddress() <= Digit7.getAddress());
         this.registerAddress = registerAddress;
@@ -18,12 +18,12 @@ public class DigitRegister implements Register {
     }
 
     @Override
-    public byte getAddress() {
+    public short getAddress() {
         return registerAddress.getAddress();
     }
 
     @Override
-    public byte getData() {
+    public short getData() {
         return data;
     }
 }
