+++ src/main/java/com/buransky/max7219/register/ShutdownRegister.java
@@ -6,23 +6,23 @@ import com.buransky.max7219.Register;
  * Table 3. Shutdown Register Format (Address (Hex) = 0xXC)
  */
 public enum ShutdownRegister implements Register {
-    ShutdownMode((byte)0x00),
-    NormalOperation((byte)0x01);
+    ShutdownMode((short)0x00),
+    NormalOperation((short)0x01);
 
     private static final RegisterAddress registerAddress = RegisterAddress.Shutdown;
-    private final byte data;
+    private final short data;
 
-    ShutdownRegister(final byte data) {
+    ShutdownRegister(final short data) {
         this.data = data;
     }
 
     @Override
-    public byte getAddress() {
+    public short getAddress() {
         return registerAddress.getAddress();
     }
 
     @Override
-    public byte getData() {
+    public short getData() {
         return data;
     }
 }
